pyproject.toml
[project]
name = "hanjan-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.11"
dependencies = [
    "anyio==4.9.0",
    "asgiref==3.9.1",
    "black>=25.1.0",
    "boto3==1.35.99",
    "certifi==2025.7.14",
    "click==8.2.1",
    "coverage>=7.9.2",
    "django==5.2.4",
    "django-cors-headers>=4.7.0",
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-redis>=6.0.0",
    "django-stubs[compatible-mypy]>=5.2.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httptools==0.6.4",
    "idna==3.10",
    "ipython>=9.4.0",
    "isort>=6.0.1",
    "mypy>=1.17.0",
    "packaging==25.0",
    "pillow>=11.3.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "requests>=2.32.4",
    "sentry-sdk[django]==2.33.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "types-requests>=2.32.4.20250611",
    "typing-extensions==4.14.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
disable_error_code = ["var-annotated"]

[[tool.mypy.overrides]]
module = "apps.*.models"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.black]
line-length = 120

[tool.isort]
profile="black"
